+++ src/main/scala/net/ddns/akgunter/rl4j_practice/environment/BanditTeamState.scala
@@ -6,7 +6,7 @@ import org.deeplearning4j.rl4j.space.Encodable
 import org.nd4j.linalg.api.ndarray.INDArray
 import org.nd4j.linalg.factory.Nd4j
 
-class BanditTeamState(curBandit: Int, rewards: Array[Array[Int]]) extends Encodable {
+class BanditTeamState(curBandit: Int, rewards: Array[Array[Double]]) extends Encodable {
   val numBandits: Int = rewards.length
   val numMachines: Int = rewards(curBandit).length
   val bestMachines: Array[Int] = BanditTeamState.generateBestRewards(rewards)
@@ -15,7 +15,7 @@ class BanditTeamState(curBandit: Int, rewards: Array[Array[Int]]) extends Encoda
 
   def getNumBandits: Int = numBandits
 
-  def getRewards: Array[Array[Int]] = rewards
+  def getRewards: Array[Array[Double]] = rewards
 
   def getBestMachines: Array[Int] = bestMachines
 
@@ -46,11 +46,11 @@ object BanditTeamState {
     new BanditTeamState(Random.nextInt(numBandits), generateRewards(numBandits, numMachines))
   }
 
-  def generateRewards(numBandits: Int, numMachines: Int): Array[Array[Int]] = {
+  def generateRewards(numBandits: Int, numMachines: Int): Array[Array[Double]] = {
     Array.fill(numBandits)(Array.fill(numMachines)(Random.nextDouble * 2 - 1))
   }
 
-  def generateBestRewards(rewards: Array[Array[Int]]): Array[Int] = {
+  def generateBestRewards(rewards: Array[Array[Double]]): Array[Int] = {
     rewards.map(_.zipWithIndex.max._2)
   }
 }
